Html related things moved to its own package
