+++ core/src/main/kotlin/renderers/html/HtmlRenderer.kt
@@ -1,10 +1,12 @@
-package org.jetbrains.dokka.renderers
+package org.jetbrains.dokka.renderers.html
 
 import kotlinx.html.*
 import kotlinx.html.stream.createHTML
 import org.jetbrains.dokka.model.Function
 import org.jetbrains.dokka.pages.*
 import org.jetbrains.dokka.plugability.DokkaContext
+import org.jetbrains.dokka.renderers.DefaultRenderer
+import org.jetbrains.dokka.renderers.OutputWriter
 import java.io.File
 
 open class HtmlRenderer(
